django_charts/requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
